chore: dissolve some standalone test files into existings files
