--- __tests__/html-block-parser.test.js
@@ -1,11 +1,0 @@
-import { mdast } from '../index';
-
-describe('Parse html block', () => {
-  it('parses an html block', () => {
-    const text = `
-<div>Some block html</div>
-    `;
-
-    expect(mdast(text)).toMatchSnapshot();
-  });
-});
